+++ photons_control/script.py
@@ -156,6 +156,220 @@ class Pipeline(object):
         if do_raise and error_catcher:
             raise RunErrors(_errors=list(set(error_catcher)))
 
+class FromGenerator(object):
+    """
+    FromGenerator let's you determine what messages to send in an async generator.
+
+    For example:
+
+    .. code-block:: python
+
+        async def gen(reference, afr, **kwargs):
+            get_power = DeviceMessages.GetPower()
+
+            for pkt, _, _ in afr.transport_target.script(get_power).run_with(reference, afr, **kwargs):
+                if pkt | DeviceMessages.StatePower:
+                    if pkt.level == 0:
+                        yield DeviceMessages.SetPower(level=65535, target=pkt.serial)
+                    else:
+                        yield DeviceMessages.SetPower(level=0, target=pkt.serial)
+
+        msg = FromGenerator(gen)
+        await target.script(msg).run_with_all("d073d5000001")
+
+    Note that all messages you yield from the generator must already have target
+    set on them.
+
+    The generator function receives the reference supplied to the run_with as well
+    as the afr object and any keyword arguments that were passed into run_with.
+
+    All messages that are yielded from the generator are sent in parallel
+
+    The return value from yield will be a future that resolves to True if the
+    message(s) were sent without error, and False if they were sent with error.
+    """
+    name = "from_generator"
+    has_children = True
+
+    def __init__(self, generator, *, reference_override=None):
+        self.generator = generator
+        self.reference_override = reference_override
+
+    def simplified(self, simplifier, chain=None):
+        return self.Item(simplifier, self.generator, self.Runner, self.reference_override)
+
+    class Item:
+        def __init__(self, simplifier, generator, runner_kls, reference_override):
+            self.generator = generator
+            self.runner_kls = runner_kls
+            self.simplifier = simplifier
+            self.reference_override = reference_override
+
+        async def run_with(self, reference, args_for_run, **kwargs):
+            do_raise = kwargs.get("error_catcher") is None
+            error_catcher = [] if do_raise else kwargs["error_catcher"]
+            kwargs["error_catcher"] = error_catcher
+
+            runner = self.runner_kls(self, reference, args_for_run, kwargs)
+
+            try:
+                async for result in runner:
+                    yield result
+            finally:
+                await runner.finish()
+
+            if do_raise and error_catcher:
+                raise RunErrors(_errors=list(set(error_catcher)))
+
+    class Runner:
+        class Done:
+            pass
+
+        def __init__(self, item, reference, args_for_run, kwargs):
+            self.item = item
+            self.kwargs = kwargs
+            self.stop_fut = asyncio.Future()
+            self.reference = reference
+            self.args_for_run = args_for_run
+            self.error_catcher = kwargs["error_catcher"]
+
+            self.ts = []
+            self.queue = asyncio.Queue()
+
+        @property
+        def generator_reference(self):
+            if self.item.reference_override in (None, True):
+                return self.reference
+            return self.item.reference_override
+
+        @property
+        def run_with_reference(self):
+            if self.item.reference_override is True:
+                return self.reference
+            elif self.item.reference_override is not None:
+                return self.item.reference_override
+
+        async def wait_for_ts(self):
+            for t in self.ts:
+                if not t.done():
+                    try:
+                        await t
+                    except asyncio.CancelledError:
+                        pass
+                    except Exception as error:
+                        hp.add_error(self.error_catcher, error)
+
+        async def finish(self):
+            await self.wait_for_ts()
+
+            self.stop_fut.cancel()
+
+            try:
+                await self.queue.put(self.Done)
+            except:
+                pass
+
+            await self._stop_ts()
+
+        def __aiter__(self):
+            self.getter_t = hp.async_as_background(self.getter())
+
+            def on_finish(res):
+                hp.async_as_background(self.queue.put(self.Done))
+            self.getter_t.add_done_callback(on_finish)
+
+            return self
+
+        async def __anext__(self):
+            nxt = await self.queue.get()
+
+            if nxt is self.Done or self.stop_fut.done():
+                raise StopAsyncIteration
+
+            return nxt
+
+        async def _stop_ts(self):
+            try:
+                if hasattr(self, "getter_t"):
+                    await self.getter_t
+            except asyncio.CancelledError:
+                pass
+            except Exception as error:
+                hp.add_error(self.error_catcher, error)
+                return
+
+        async def getter(self):
+            gen = self.item.generator(self.generator_reference, self.args_for_run, **self.kwargs)
+            complete = None
+
+            while True:
+                try:
+                    msg = await gen.asend(complete)
+                    if isinstance(msg, Exception):
+                        hp.add_error(self.error_catcher, msg)
+                        continue
+
+                    if self.stop_fut.done():
+                        break
+
+                    complete = asyncio.Future()
+
+                    f_for_items = []
+                    for item in self.item.simplifier(msg):
+                        f = asyncio.Future()
+                        f_for_items.append(f)
+                        t = hp.async_as_background(self.retrieve(item, f))
+                        self.ts.append(t)
+
+                    self.complete_on_all_done(f_for_items, complete)
+                    self.ts = [t for t in self.ts if not t.done()]
+                except StopAsyncIteration:
+                    break
+
+            await self.wait_for_ts()
+
+        async def retrieve(self, item, f):
+            i = {"success": True}
+
+            def pass_on_error(e):
+                i["success"] = False
+                hp.add_error(self.error_catcher, e)
+
+            kwargs = dict(self.kwargs)
+            kwargs["error_catcher"] = pass_on_error
+
+            try:
+                async for info in item.run_with(self.run_with_reference, self.args_for_run, **kwargs):
+                    await self.queue.put(info)
+            finally:
+                if not f.done():
+                    f.set_result(i["success"])
+
+        def complete_on_all_done(self, fs, complete):
+            def finish(res):
+                if complete.done():
+                    return
+
+                if res.cancelled():
+                    complete.cancel()
+
+                for f in fs:
+                    if f.cancelled():
+                        complete.cancel()
+                        return
+
+                    exc = f.exception()
+                    if exc is not None:
+                        complete.set_result(False)
+                        return
+                    elif f.result() is False:
+                        complete.set_result(False)
+                        return
+
+                complete.set_result(True)
+            waiter = hp.async_as_background(asyncio.wait(fs))
+            waiter.add_done_callback(finish)
+
 class Decider(object):
     """
     This one is a complex beast used for determining what to do based on the